+++ client/components/HeaderBar/styles.js
@@ -1,0 +1,25 @@
+const styles = {
+  headerBar: {
+    backgroundColor: '#4a90e2'
+  },
+  backButton: {
+    marginRight: 20,
+    color: '#ffffff'
+  },
+  action: {
+    flex: 1,
+    fontSize: 20,
+    letterSpacing: 1.4,
+    textTransform: 'uppercase',
+    textAlign: 'left',
+    color: '#ffffff'
+  },
+  title: {
+    fontFamily: 'Barrio',
+    fontSize: 36,
+    letterSpacing: 2.6,
+    color: '#ffffff'
+  }
+};
+
+export default styles;
