+++ client/scenes/PublishProject/styles.js
@@ -1,0 +1,92 @@
+const styles = {
+  root: {
+    flexGrow: 1,
+    maxWidth: 1280,
+    margin: '20px auto',
+    backgroundColor: '#f8f9fc;'
+  },
+  flex: {
+    display: 'flex'
+  },
+  flex1: {
+    flex: 1
+  },
+  hidden: {
+    display: 'none'
+  },
+  fillParent: {
+    width: '100%',
+    height: '100%'
+  },
+  form: {
+    padding: 20,
+    backgroundColor: '#ffffff;'
+  },
+  formRow: {
+    marginTop: 30,
+    marginBottom: 40
+  },
+  formItem3Col: {
+    paddingRight: 30,
+    width: '25%'
+  },
+  formItem2Col: {
+    paddingRight: 30,
+    width: '50%'
+  },
+  counter: {
+    textAlign: 'right'
+  },
+  projectPhoto: {
+    width: 160,
+    height: 160,
+    color: '#9e9e9e'
+  },
+  boxProjectPhoto: {
+    width: '30%',
+    height: '100%',
+    textAlign: 'center',
+    padding: '60px 60px',
+    margin: '0 auto',
+    border: 'solid 1px #979797',
+    backgroundColor: '#D6D6D6'
+  },
+  labelProjectPhoto: {
+    borderBottom: 'solid 1px #ffffff',
+    color: '#ffffff'
+  },
+  boxIcon: {
+    alignItems: 'center',
+    padding: 10,
+    marginRight: 30,
+    width: 120
+  },
+  boxIconSelected: {
+    fontWeight: 'bold',
+    border: 'solid 1px #979797',
+    backgroundColor: '#D6D6D6'
+  },
+  icon: {
+    width: 60,
+    height: 60,
+    color: '#9e9e9e'
+  },
+  textIcon: {
+    paddingTop: 9
+  },
+  link: {
+    fontFamily: 'Roboto',
+    fontSize: 14,
+    fontWeight: 500,
+    padding: '10px 15px',
+    textTransform: 'uppercase',
+    textDecoration: 'none',
+    color: '#ffffff',
+    letterSpacing: 0.5,
+    backgroundColor: '#4a90e2',
+    borderRadius: 2,
+    boxShadow: '0 2px 2px 0 rgba(0, 0, 0, 0.24), 0 0 2px 0 rgba(0, 0, 0, 0.12)'
+  }
+};
+
+export default styles;
